feat(frontend): 类型、常量、枚举优化调整 #5151 # Reviewed, transaction id: 10524
