+++ dbm-ui/frontend/src/common/const/accountTypes.ts
@@ -1,0 +1,10 @@
+/**
+ * 账号类型
+ */
+export enum AccountTypes {
+  MYSQL = 'mysql',
+  TENDBCLUSTER = 'tendbcluster',
+  MONGODB = 'mongodb',
+  SQLSERVER = 'sqlserver',
+}
+export type AccountTypesValues = `${AccountTypes}`;
