+++ dbm-ui/frontend/src/common/const/dbSysExclude.ts
@@ -1,0 +1,11 @@
+/**
+ * 要排除的系统库名称
+ */
+export const dbSysExclude = [
+  'mysql',
+  'db_infobase',
+  'information_schema',
+  'performance_schema',
+  'sys',
+  'infodba_schema',
+];
