+++ dbm-ui/frontend/src/common/const/dbTypes.ts
@@ -1,0 +1,25 @@
+/**
+ * 数据库类型
+ */
+export enum DBTypes {
+  MYSQL = 'mysql',
+  TENDBCLUSTER = 'tendbcluster', // TenDBCluster
+}
+export enum DBTypes {
+  REDIS = 'redis',
+}
+export enum DBTypes {
+  ES = 'es', // ElasticSearch
+  KAFKA = 'kafka',
+  HDFS = 'hdfs',
+  INFLUXDB = 'influxdb',
+  RIAK = 'riak',
+  PULSAR = 'pulsar',
+  DORIS = 'doris',
+}
+export enum DBTypes {
+  MONGODB = 'mongodb',
+}
+export enum DBTypes {
+  SQLSERVER = 'sqlserver',
+}
