+++ dbm-ui/frontend/src/common/const/localStorageKeys.ts
@@ -1,0 +1,6 @@
+export enum LocalStorageKeys {
+  REDIS_ROLLBACK_LIST = 'REDIS_ROLLBACK_LIST', // 跨页使用的回写数据列表
+  REDIS_DB_REPLACE_MASTER_TIP = 'REDIS_DB_REPLACE_MASTER_TIP', // 整机替换弹窗中的master提示
+  REDIS_DATA_CHECK_AND_REPAIR = 'REDIS_DATA_CHECK_AND_REPAIR', // 跨页使用的数据检验与修复的行数据
+  REDIS_DB_DATA_RECORD_RECOPY = 'REDIS_DB_DATA_RECORD_RECOPY', // 跨页使用的数据复制记录
+}
