+++ dbm-ui/frontend/src/common/const/machineTypes.ts
@@ -1,0 +1,47 @@
+/**
+ * 机器类型
+ */
+export enum MachineTypes {
+  SPIDER = 'spider',
+  REMOTE = 'remote',
+  PROXY = 'proxy',
+  BACKEND = 'backend',
+  SINGLE = 'single',
+  PREDIXY = 'predixy',
+  TWEMPROXY = 'twemproxy',
+  INFLUXDB = 'influxdb',
+  RIAK = 'riak',
+}
+export enum MachineTypes {
+  REDIS = 'redis',
+  TENDISCACHE = 'tendiscache',
+  TENDISSSD = 'tendisssd',
+  TENDISPLUS = 'tendisplus',
+}
+export enum MachineTypes {
+  ES_DATANODE = 'es_datanode',
+  ES_MASTER = 'es_master',
+  ES_CLIENT = 'es_client',
+}
+export enum MachineTypes {
+  HDFS_MASTER = 'hdfs_master',
+  HDFS_DATANODE = 'hdfs_datanode',
+}
+export enum MachineTypes {
+  MONGOS = 'mongos',
+  MONGODB = 'mongodb',
+  MONGO_CONFIG = 'mongo_config',
+}
+export enum MachineTypes {
+  SQLSERVER_HA = 'sqlserver_ha',
+  SQLSERVER_SINGLE = 'sqlserver_single',
+}
+export enum MachineTypes {
+  ZOOKEEPER = 'zookeeper',
+  PULSAR_ZOOKEEPER = 'pulsar_zookeeper',
+  PULSAR_BOOKKEEPER = 'pulsar_bookkeeper',
+}
+export enum MachineTypes {
+  BROKER = 'broker',
+  PULSAR_BROKER = 'pulsar_broker',
+}
