+++ dbm-ui/frontend/src/common/const/occupiedInnerHeight.ts
@@ -1,0 +1,8 @@
+/**
+ * 用于列表界面计算table最大高度（该值包括导航/面包屑/容器边距/表格过滤操作高度）
+ */
+export enum OccupiedInnerHeight {
+  WITH_PAGINATION = 218,
+  NOT_PAGINATION = 202,
+  NOT_PAGINATION_WITH_TIP = 284,
+}
