+++ dbm-ui/frontend/src/common/const/osTypes.ts
@@ -1,0 +1,12 @@
+/**
+ * 操作系统类型
+ */
+export enum OSTypes {
+  Linux = 1,
+  Windows = 2,
+  AIX = 3,
+  Unix = 4,
+  Solaris = 5,
+  FreeBSD = 7,
+  MacOS = 8,
+}
