+++ dbm-ui/frontend/src/components/cluster-tab/Index.vue
@@ -14,7 +14,7 @@
 <template>
   <BkTab
     v-model:active="moduleValue"
-    class="top-tabs"
+    class="cluster-tab"
     type="unborder-card">
     <BkTabPanel
       v-for="tab of renderTabs"
@@ -25,8 +25,6 @@
 </template>
 
 <script setup lang="ts">
-  import { useI18n } from 'vue-i18n';
-
   import type {
     ControllerBaseInfo,
     ExtractedControllerDataKeys,
@@ -34,151 +32,12 @@
 
   import { useFunController } from '@stores';
 
-  import { ClusterTypes } from '@common/const';
+  import { clusterTypeInfos, ClusterTypes } from '@common/const';
 
   interface TabItem {
-    moduleId: ExtractedControllerDataKeys;
     label: string;
     name: ClusterTypes;
   }
-  const { t } = useI18n();
-  const tabs: TabItem[] = [
-    {
-      moduleId: 'mysql',
-      label: t('MySQL 单节点'),
-      name: ClusterTypes.TENDBSINGLE,
-    },
-    {
-      moduleId: 'mysql',
-      label: t('MySQL 主从'),
-      name: ClusterTypes.TENDBHA,
-    },
-    {
-      moduleId: 'mysql',
-      label: 'TenDBCluster',
-      name: ClusterTypes.TENDBCLUSTER,
-    },
-    {
-      moduleId: 'redis',
-      label: 'TendisplusCluster',
-      name: ClusterTypes.TENDIS_PLUS_CLUSTER,
-    },
-    {
-      moduleId: 'redis',
-      label: 'TendisplusInstance',
-      name: ClusterTypes.TENDIS_PLUS_INSTANCE,
-    },
-    {
-      moduleId: 'redis',
-      label: 'TendisSSDInstance',
-      name: ClusterTypes.TENDIS_SSD_INSTANCE,
-    },
-    // {
-    //   moduleId: 'redis',
-    //   label: 'Redis',
-    //   name: ClusterTypes.REDIS,
-    // },
-    {
-      moduleId: 'redis',
-      label: 'RedisCluster',
-      name: ClusterTypes.REDIS_CLUSTER,
-    },
-    {
-      moduleId: 'redis',
-      label: t('Redis 主从'),
-      name: ClusterTypes.REDIS_INSTANCE,
-    },
-    {
-      moduleId: 'redis',
-      label: t('Redis 分片'),
-      name: ClusterTypes.PREDIXY_REDIS_CLUSTER,
-    },
-    {
-      moduleId: 'redis',
-      label: 'Tendisplus',
-      name: ClusterTypes.PREDIXY_TENDISPLUS_CLUSTER,
-    },
-    {
-      moduleId: 'redis',
-      label: 'Redis',
-      name: ClusterTypes.TWEMPROXY_REDIS_INSTANCE,
-    },
-    {
-      moduleId: 'redis',
-      label: 'Twemproxy',
-      name: ClusterTypes.TWEMPROXY_TENDISPLUS_INSTANCE,
-    },
-    {
-      moduleId: 'redis',
-      label: 'TendisSSD',
-      name: ClusterTypes.TWEMPROXY_TENDIS_SSD_INSTANCE,
-    },
-    // {
-    //   moduleId: 'mongodb',
-    //   label: 'MongoDB',
-    //   name: ClusterTypes.MONGODB,
-    // },
-    {
-      moduleId: 'mongodb',
-      label: 'MongoCluster',
-      name: ClusterTypes.MONGOCLUSTER,
-    },
-    {
-      moduleId: 'mongodb',
-      label: t('MongoDB 副本集'),
-      name: ClusterTypes.MONGO_REPLICA_SET,
-    },
-    {
-      moduleId: 'mongodb',
-      label: t('MongoDB 分片集'),
-      name: ClusterTypes.MONGO_SHARED_CLUSTER,
-    },
-    {
-      moduleId: 'sqlserver',
-      label: t('SQLServer 单节点'),
-      name: ClusterTypes.SQLSERVER_SINGLE,
-    },
-    {
-      moduleId: 'sqlserver',
-      label: t('SQLServer 主从'),
-      name: ClusterTypes.SQLSERVER_HA,
-    },
-    {
-      moduleId: 'bigdata',
-      label: 'ES',
-      name: ClusterTypes.ES,
-    },
-    {
-      moduleId: 'bigdata',
-      label: 'Kafka',
-      name: ClusterTypes.KAFKA,
-    },
-    {
-      moduleId: 'bigdata',
-      label: 'HDFS',
-      name: ClusterTypes.HDFS,
-    },
-    {
-      moduleId: 'bigdata',
-      label: 'Pulsar',
-      name: ClusterTypes.PULSAE,
-    },
-    {
-      moduleId: 'bigdata',
-      label: 'InfluxDB',
-      name: ClusterTypes.INFLUXDB,
-    },
-    {
-      moduleId: 'bigdata',
-      label: 'Riak',
-      name: ClusterTypes.RIAK,
-    },
-    {
-      moduleId: 'bigdata',
-      label: 'Doris',
-      name: ClusterTypes.DORIS,
-    },
-  ];
 
   const funControllerStore = useFunController();
 
@@ -186,23 +45,30 @@
     default: ClusterTypes.TENDBSINGLE,
   });
 
-  const renderTabs = tabs.filter((item) => {
-    const { moduleId, name: clusterType } = item;
-    const data = funControllerStore.funControllerData[moduleId];
-    // 整个模块没有开启
-    if (!data || data.is_enabled !== true) {
-      return false;
+  const renderTabs = Object.values(clusterTypeInfos).reduce((result, item) => {
+    const { id: clusterType, name, dbType, moduleId } = item;
+    const realModuleId = moduleId ?? (dbType as ExtractedControllerDataKeys);
+    const data = funControllerStore.funControllerData[realModuleId];
+    if (dbType === moduleId && data?.is_enabled) {
+      result.push({
+        label: name,
+        name: clusterType,
+      });
+    } else {
+      const children = data?.children as Record<ClusterTypes, ControllerBaseInfo>;
+      if (children[clusterType]?.is_enabled) {
+        result.push({
+          label: name,
+          name: clusterType,
+        });
+      }
     }
-    if (clusterType === item.moduleId && data.is_enabled) {
-      return true;
-    }
-    const children = data.children as Record<ClusterTypes, ControllerBaseInfo>;
-    return children[clusterType] && children[clusterType]?.is_enabled;
-  });
+    return result;
+  }, [] as TabItem[]);
 </script>
 
 <style lang="less">
-  .top-tabs {
+  .cluster-tab {
     padding: 0 24px;
     background: #fff;
     box-shadow: 0 3px 4px 0 rgb(0 0 0 / 4%);
