+++ dbm-ui/frontend/src/locales/zh-cn.json
@@ -3149,6 +3149,26 @@
     "较目标容量超出n台": "，较目标容量超出 {0} 台",
     "Observer类型节点若存在至少保留两台": "Observer类型节点若存在至少保留两台",
     "请输入 where 条件，如：userId > 10000，不要带where关键字": "请输入 where 条件，如：userId > 10000，不要带where关键字",
+    "Tendisplus存储版集群": "Tendisplus存储版集群",
+    "TendisCache集群": "TendisCache集群",
+    "Tendis存储版集群": "Tendis存储版集群",
+    "RedisCache主从版": "RedisCache主从版",
+    "TendisSSD主从版": "TendisSSD主从版",
+    "Tendisplus主从版": "Tendisplus主从版",
+    "RedisCluster集群": "RedisCluster集群",
+    "TendisplusCluster集群": "TendisplusCluster集群",
+    "Mongo副本集群": "Mongo副本集群",
+    "Mongo分片集群": "Mongo分片集群",
+    "SQLServer单节点版": "SQLServer单节点版",
+    "SQLServer主从版": "SQLServer主从版",
+    "MySQL 单节点": "MySQL 单节点",
+    "MySQL 主从": "MySQL 主从",
+    "Redis 分片": "Redis 分片",
+    "MongoDB 副本集": "MongoDB 副本集",
+    "MongoDB 分片集": "MongoDB 分片集",
+    "SQLServer 单节点": "SQLServer 单节点",
+    "SQLServer 主从": "SQLServer 主从",
+    "备份工具-TXSQL": "备份工具-TXSQL",
     "范围n_min_max": "范围 {n}, {min} ~ {max}",
     "这行勿动！新增翻译请在上一行添加！": ""
 }
