+++ dbm-ui/frontend/src/views/db-configure/business/list/components/hooks/useTreeData.ts
@@ -190,7 +190,7 @@ export const useTreeData = (treeState: TreeState) => {
           ClusterTypes.KAFKA,
           ClusterTypes.HDFS,
           ClusterTypes.INFLUXDB,
-          ClusterTypes.PULSAE,
+          ClusterTypes.PULSAR,
           ClusterTypes.RIAK,
           ClusterTypes.DORIS,
         ].includes(value);
