+++ dbm-ui/frontend/src/views/db-configure/hooks/useLevelParams.ts
@@ -9,26 +9,21 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed
  * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
  * the specific language governing permissions and limitations under the License.
-*/
+ */
 
 import type { ComputedRef } from 'vue';
 
 import { useGlobalBizs } from '@stores';
 
-import {
-  clusterTypeInfos,
-  type ClusterTypesValues,
-  ConfLevels,
-  type ConfLevelValues,
-} from '@common/const';
+import { clusterTypeInfos, ClusterTypes, ConfLevels, type ConfLevelValues } from '@common/const';
 
 import { notModuleClusters } from '../common/const';
 
 export type LevelParams = {
-  level_name?: ConfLevelValues,
-  level_value?: number,
-  level_info?: any,
-  bk_biz_id?: number
+  level_name?: ConfLevelValues;
+  level_value?: number;
+  level_info?: any;
+  bk_biz_id?: number;
 };
 
 /**
@@ -40,12 +35,19 @@ export const useLevelParams = (isPlat: boolean): ComputedRef<LevelParams> => {
 
   return computed(() => {
     // 平台层级不需要额外参数
-    if (isPlat) return {};
+
+    // 平台层级不需要额外参数
+    if (isPlat) {
+      return {};
+    }
 
     const { treeId, parentId, clusterType } = route.params;
-    if (!treeId) return {
-      bk_biz_id: globalBizsStore.currentBizId,
-    };
+
+    if (!treeId) {
+      return {
+        bk_biz_id: globalBizsStore.currentBizId,
+      };
+    }
 
     const params = {
       level_name: ConfLevels.PLAT,
@@ -57,7 +59,7 @@ export const useLevelParams = (isPlat: boolean): ComputedRef<LevelParams> => {
     const [levelType, nodeId] = (treeId as string).split('-');
     params.level_name = levelType as ConfLevels;
     params.level_value = Number(nodeId);
-    const notExistModule = notModuleClusters.includes(clusterTypeInfos[clusterType as ClusterTypesValues].dbType);
+    const notExistModule = notModuleClusters.includes(clusterTypeInfos[clusterType as ClusterTypes].dbType);
     if (parentId && levelType === ConfLevels.CLUSTER) {
       let [parentLevelType, parentNodeId] = (parentId as string).split('-');
       if (notExistModule) {
