+++ dbm-ui/frontend/src/views/pulsar-manage/list/components/detail/components/node-list/Index.vue
@@ -268,7 +268,7 @@
     clearSearchValue,
     validateSearchValues,
     handleSearchValueChange,
-  } = useLinkQueryColumnSerach(ClusterTypes.PULSAE, [
+  } = useLinkQueryColumnSerach(ClusterTypes.PULSAR, [
     'bk_cloud_id',
   ], () => fetchNodeList());
